+++ packages/core-js/proposals/symbol-predicates-v2.js
@@ -1,0 +1,3 @@
+// https://github.com/tc39/proposal-symbol-predicates
+require('../modules/esnext.symbol.is-registered-symbol');
+require('../modules/esnext.symbol.is-well-known-symbol');